requirements.txt
# Unsloth (chọn đúng version theo môi trường)
unsloth[kaggle-new] @ git+https://github.com/unslothai/unsloth.git

# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.14.0
trl>=0.8.0
peft>=0.10.0
accelerate>=0.27.0
bitsandbytes>=0.43.0

# Utils
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
